merge from original apache/tomcat repository

+++ java/org/apache/catalina/tribes/group/interceptors/ThroughputInterceptor.java
@@ -5,9 +5,9 @@
  * The ASF licenses this file to You under the Apache License, Version 2.0
  * (the "License"); you may not use this file except in compliance with
  * the License.  You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -30,14 +30,6 @@ import org.apache.catalina.tribes.io.XByteBuffer;
 import org.apache.juli.logging.Log;
 import org.apache.juli.logging.LogFactory;
 
-
-
-/**
- *
- *
- * @author Filip Hanik
- * @version 1.0
- */
 public class ThroughputInterceptor extends ChannelInterceptorBase {
     private static final Log log = LogFactory.getLog(ThroughputInterceptor.class);
 
@@ -66,13 +58,13 @@ public class ThroughputInterceptor extends ChannelInterceptorBase {
             msgTxErr.addAndGet(1);
             if ( access.get() == 1 ) access.addAndGet(-1);
             throw x;
-        } 
+        }
         mbTx += (bytes*destination.length)/(1024d*1024d);
         mbAppTx += bytes/(1024d*1024d);
         if ( access.addAndGet(-1) == 0 ) {
             long stop = System.currentTimeMillis();
             timeTx += (stop - txStart) / 1000d;
-            if ((msgTxCnt.get() / interval) >= lastCnt) {
+            if ((msgTxCnt.get() / (double) interval) >= lastCnt) {
                 lastCnt++;
                 report(timeTx);
             }
@@ -88,9 +80,9 @@ public class ThroughputInterceptor extends ChannelInterceptorBase {
         msgRxCnt.addAndGet(1);
         if ( msgRxCnt.get() % interval == 0 ) report(timeTx);
         super.messageReceived(msg);
-        
+
     }
-    
+
     public void report(double timeTx) {
         StringBuilder buf = new StringBuilder("ThroughputInterceptor Report[\n\tTx Msg:");
         buf.append(msgTxCnt).append(" messages\n\tSent:");
@@ -112,7 +104,7 @@ public class ThroughputInterceptor extends ChannelInterceptorBase {
         buf.append(df.format(mbRx)).append(" MB]\n");
         if ( log.isInfoEnabled() ) log.info(buf);
     }
-    
+
     public void setInterval(int interval) {
         this.interval = interval;
     }
